# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DealPilot Local API
  version: 1.0.0
  description: |
    DealPilot 本地优先架构的后端 API 契约。
    所有端点前缀 /api/v1，除 /health 外全部需要 Bearer Token 认证。
    Agent 只监听 127.0.0.1，禁止 0.0.0.0。
    列表响应使用游标分页（cursor + next_cursor）。
    写操作支持 Idempotency-Key 请求头实现幂等重试。
  contact:
    name: DealPilot
servers:
  - url: http://127.0.0.1:{port}
    description: 本地 Agent 回环地址
...
paths:
...
  /api/v1/customers:
    get:
      tags: [customers]
      summary: 客户列表（游标分页、搜索、筛选、排序）
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search
          in: query
          description: 按名称/公司/国家模糊搜索
          schema: { type: string }
        - name: grade
          in: query
          description: 客户分级筛选 A/B/C
          schema: { $ref: '#/components/schemas/CustomerGrade' }
        - name: status
          in: query
          description: 客户状态筛选
          schema: { type: string }
        - name: sort
          in: query
          description: 排序字段（created_at, name, grade 等）
          schema: { type: string, default: created_at }
      responses:
        '200':
          description: 客户列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [customers]
      summary: 新建客户
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerCreate' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/customers/{id}:
    get:
      tags: [customers]
      summary: 客户档案详情
      description: 含联系人、社媒、跟进时间线、项目、未完成提醒。
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 客户详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomerDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      tags: [customers]
      summary: 编辑客户
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerUpdate' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [customers]
      summary: 软删除客户（30天可恢复）
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      responses:
        '204':
          description: 软删除成功
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/customers/{id}/restore:
    post:
      tags: [customers]
      summary: 恢复已删除客户
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      responses:
        '200':
          description: 恢复成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/customers/merge:
    post:
      tags: [customers]
      summary: 合并客户（跟进/项目/提醒/社媒迁移）
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerMergeRequest' }
      responses:
        '200':
          description: 合并成功，返回保留客户
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Customer' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/customers/{id}/contacts:
    get:
      tags: [customers]
      summary: 联系人列表
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: 联系人列表
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Contact' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [customers]
      summary: 新建联系人
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactCreate' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contact' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/contacts/{id}:
    put:
      tags: [customers]
      summary: 编辑联系人
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactUpdate' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Contact' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [customers]
      summary: 删除联系人
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      responses:
        '204':
          description: 删除成功
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/v1/backups/create:
    post:
      tags: [backups]
      summary: 创建加密备份
      description: Argon2id 派生密钥 + AES-256-GCM 加密。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BackupCreateRequest' }
      responses:
        '200':
          description: 加密备份文件下载
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/backups/validate:
    post:
      tags: [backups]
      summary: 校验备份完整性和兼容性
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                password:
                  type: string
                  format: password
              required: [file, password]
      responses:
        '200':
          description: 校验结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BackupValidateResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/backups/restore:
    post:
      tags: [backups]
      summary: 恢复备份（全局互斥锁）
      description: 恢复期间拒绝其他写请求；失败不覆盖当前数据库。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                password:
                  type: string
                  format: password
              required: [file, password]
      responses:
        '200':
          description: 恢复结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BackupRestoreResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
    CursorParam:
      name: cursor
      in: query
      description: 游标分页位置
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      description: 每页数量
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      description: 幂等键，重复请求返回首次结果
      schema: { type: string }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 短期访问令牌，工作台来自 sessionStorage，插件来自 Native Messaging Bootstrap。
  responses:
    BadRequest:
      description: 请求参数或格式错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: 未认证或令牌无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Origin 校验失败或权限不足
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 资源冲突或状态非法
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: 错误码
        message:
          type: string
          description: 错误信息
        details:
          type: object
          description: 附加详情
        request_id:
          type: string
          description: 请求追踪 ID
      required: [code, message, request_id]
...
    CustomerGrade:
      type: string
      enum: [A, B, C]
    CustomerStatus:
      type: string
      enum: [active, inactive]
    Customer:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        company: { type: string, nullable: true }
        country: { type: string, nullable: true }
        source: { type: string, nullable: true }
        grade: { $ref: '#/components/schemas/CustomerGrade' }
        status: { $ref: '#/components/schemas/CustomerStatus' }
        deleted_at: { type: string, nullable: true, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, name, grade, status, created_at, updated_at]
    CustomerCreate:
      type: object
      properties:
        name: { type: string }
        company: { type: string }
        country: { type: string }
        source: { type: string }
        grade: { $ref: '#/components/schemas/CustomerGrade' }
        status: { $ref: '#/components/schemas/CustomerStatus' }
      required: [name, grade]
    CustomerUpdate:
      type: object
      properties:
        name: { type: string }
        company: { type: string }
        country: { type: string }
        source: { type: string }
        grade: { $ref: '#/components/schemas/CustomerGrade' }
        status: { $ref: '#/components/schemas/CustomerStatus' }
    CustomerDetail:
      type: object
      allOf:
        - $ref: '#/components/schemas/Customer'
        - type: object
          properties:
            contacts:
              type: array
              items: { $ref: '#/components/schemas/Contact' }
            social_accounts:
              type: array
              items: { $ref: '#/components/schemas/SocialAccount' }
            follow_ups:
              type: array
              items: { $ref: '#/components/schemas/FollowUp' }
            projects:
              type: array
              items: { $ref: '#/components/schemas/Project' }
            pending_reminders:
              type: array
              items: { $ref: '#/components/schemas/Reminder' }
    CustomerListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Customer' }
        next_cursor: { type: string, nullable: true }
      required: [items]
    CustomerMergeRequest:
      type: object
      properties:
        source_id: { type: string }
        target_id: { type: string }
        field_resolutions:
          type: object
          description: 冲突字段的手动裁决结果
      required: [source_id, target_id]
    Contact:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        name: { type: string }
        title: { type: string, nullable: true }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
      required: [id, customer_id, name, created_at]
    ContactCreate:
      type: object
      properties:
        name: { type: string }
        title: { type: string }
        email: { type: string }
        phone: { type: string }
      required: [name]
    ContactUpdate:
      type: object
      properties:
        name: { type: string }
        title: { type: string }
        email: { type: string }
        phone: { type: string }
    SocialAccount:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        contact_id: { type: string, nullable: true }
        platform: { type: string }
        raw_identifier: { type: string }
        normalized_identifier: { type: string }
        manually_bound: { type: boolean }
        created_at: { type: string, format: date-time }
      required: [id, customer_id, platform, raw_identifier, normalized_identifier, manually_bound, created_at]
...
    FollowUpType:
      type: string
      enum: [call, email, chat, visit, note, message]
    MessageDirection:
      type: string
      enum: [inbound, outbound]
    FollowUp:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        project_id: { type: string, nullable: true }
        type: { $ref: '#/components/schemas/FollowUpType' }
        note: { type: string, nullable: true }
        message_body: { type: string, nullable: true }
        message_direction: { $ref: '#/components/schemas/MessageDirection', nullable: true }
        occurred_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
      required: [id, customer_id, type, occurred_at, created_at]
...
    ReminderType:
      type: string
      enum: [fixed_time, waiting_reply, paused]
    ReminderStatus:
      type: string
      enum: [pending, completed, snoozed, ignored, overdue, replied]
    ReminderPriority:
      type: string
      enum: [low, normal, high, urgent]
    Reminder:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        project_id: { type: string, nullable: true }
        type: { $ref: '#/components/schemas/ReminderType' }
        status: { $ref: '#/components/schemas/ReminderStatus' }
        due_at: { type: string, format: date-time }
        priority: { $ref: '#/components/schemas/ReminderPriority' }
        last_notified_at: { type: string, nullable: true, format: date-time }
        snooze_until: { type: string, nullable: true, format: date-time }
        resolution: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, customer_id, type, status, due_at, priority, created_at, updated_at]
...
    ProjectStage:
      type: string
      enum: [lead, qualified, proposal, negotiation, closed_won, closed_lost, archived]
    ProjectGrade:
      type: string
      enum: [S, A, B, C]
    Project:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        name: { type: string }
        currency: { type: string }
        amount: { type: number, nullable: true }
        probability: { type: integer, minimum: 0, maximum: 100, nullable: true }
        expected_close_date: { type: string, nullable: true, format: date }
        stage: { $ref: '#/components/schemas/ProjectStage' }
        grade: { $ref: '#/components/schemas/ProjectGrade' }
        closed_reason: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, customer_id, name, stage, grade, created_at, updated_at]
...
    BackupCreateRequest:
      type: object
      properties:
        password: { type: string, format: password }
      required: [password]
    BackupValidateResponse:
      type: object
      properties:
        valid: { type: boolean }
        schema_version: { type: string }
        app_version: { type: string }
        integrity_ok: { type: boolean }
      required: [valid, integrity_ok]
    BackupRestoreResponse:
      type: object
      properties:
        success: { type: boolean }
        rows_restored: { type: integer }
      required: [success, rows_restored]
...
security:
  - BearerAuth: []